+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `displayAsCurrency` to `FormElement`
 - `storageType` to `FormElement`
+- `formsAppEnvironment` to `FormsClient Search`
 
 ## 2.1.0 (2021-04-27)